refactor: organiza modulo graphdb
