+++ api/src/db/graphdb/nodes.rs
@@ -1,11 +1,11 @@
-use crate::db::TryCast;
+use crate::db::graphdb::TryCast;
 use kuzu::{
     Value,
     Connection
 };
-use crate::db::QueryResultUtil;
+use crate::db::graphdb::QueryResultUtil;
 use crate::db::models::Node;
-use crate::db::ConnectionUtil;
+use crate::db::graphdb::ConnectionUtil;
 
 pub fn node_create(
     conn: &Connection<'_>,
